requirements.txt
pytube==12.1.0
Django==4.1.2
python-telegram-bot==13.14